requirements.txt
numpy~=2.2.2
matplotlib~=3.10.0
scipy~=1.15.1
scikit-learn~=1.6.1
networkx~=3.1
sympy~=1.11.1
pandas~=2.2.3
PyQt5~=5.15.11
pillow~=11.1.0
scikit-image~=0.25.1
shapely~=2.0.7